Revert "Merge branch 'liferay:master' into LPS-151442"
This reverts commit 36f042141ccf002bada910b2a2e739a1790c0946, reversing
changes made to a0a23ab39d8e3de69fe4a2dc4a74a4711fae2b81.

+++ modules/apps/layout/layout-content-page-editor-web/src/main/java/com/liferay/layout/content/page/editor/web/internal/portlet/action/DeleteSegmentsExperienceMVCActionCommand.java
@@ -26,6 +26,7 @@ import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.segments.constants.SegmentsExperienceConstants;
 import com.liferay.segments.service.SegmentsExperienceService;
 
 import java.util.List;
@@ -70,8 +71,10 @@ public class DeleteSegmentsExperienceMVCActionCommand
 		long segmentsExperienceId = ParamUtil.getLong(
 			actionRequest, "segmentsExperienceId");
 
-		_segmentsExperienceService.deleteSegmentsExperience(
-			segmentsExperienceId);
+		if (segmentsExperienceId != SegmentsExperienceConstants.ID_DEFAULT) {
+			_segmentsExperienceService.deleteSegmentsExperience(
+				segmentsExperienceId);
+		}
 
 		List<FragmentEntryLink> fragmentEntryLinks =
 			_fragmentEntryLinkLocalService.
